Added models Families, FamilyInvite and FamilyMemvers. Also updated controllers AccountController, UserController, BudgetController and CategoryController for dealing with famylies and invitations in families. Also updated views for these controllers. Now they are working with families

+++ models/Users.h
@@ -17,7 +17,7 @@
 #endif
 #include <trantor/utils/Date.h>
 #include <trantor/utils/Logger.h>
-#include <jsoncpp/json/json.h>
+#include <json/json.h>
 #include <string>
 #include <string_view>
 #include <memory>
@@ -55,7 +55,7 @@ class Users
     static const std::string tableName;
     static const bool hasPrimaryKey;
     static const std::string primaryKeyName;
-    using PrimaryKeyType = int32_t;
+    using PrimaryKeyType = int64_t;
     const PrimaryKeyType &getPrimaryKey() const;
 
     /**
@@ -102,11 +102,11 @@ class Users
 
     /**  For column id  */
     ///Get the value of the column id, returns the default value if the column is null
-    const int32_t &getValueOfId() const noexcept;
+    const int64_t &getValueOfId() const noexcept;
     ///Return a shared_ptr object pointing to the column const value, or an empty shared_ptr object if the column is null
-    const std::shared_ptr<int32_t> &getId() const noexcept;
+    const std::shared_ptr<int64_t> &getId() const noexcept;
     ///Set the value of the column id
-    void setId(const int32_t &pId) noexcept;
+    void setId(const int64_t &pId) noexcept;
 
     /**  For column name  */
     ///Get the value of the column name, returns the default value if the column is null
@@ -166,7 +166,7 @@ class Users
     void updateArgs(drogon::orm::internal::SqlBinder &binder) const;
     ///For mysql or sqlite3
     void updateId(const uint64_t id);
-    std::shared_ptr<int32_t> id_;
+    std::shared_ptr<int64_t> id_;
     std::shared_ptr<std::string> name_;
     std::shared_ptr<std::string> email_;
     std::shared_ptr<std::string> hashedPassword_;
